eigen and eth amounts stored

+++ src/contracts/interfaces/IRegistrationManager.sol
@@ -2,6 +2,6 @@
 pragma solidity ^0.8.9;
 
 interface IRegistrationManager {
-	function registerOperator(address, bytes calldata) external returns(uint8, uint256);
+	function registerOperator(address, bytes calldata) external returns(uint8, uint128);
 	function deregisterOperator(address, bytes calldata) external returns(bool);
 }
