+++ src/contracts/interfaces/IVoteWeighter.sol
@@ -2,6 +2,6 @@
 pragma solidity ^0.8.9;
 
 interface IVoteWeighter {
-	function weightOfOperatorEth(address) external returns(uint256);
-	function weightOfOperatorEigen(address) external returns(uint256);
+	function weightOfOperatorEth(address) external returns(uint128);
+	function weightOfOperatorEigen(address) external returns(uint128);
 }
